requirements.txt
contourpy==1.1.0
cycler==0.11.0
et-xmlfile==1.1.0
Faker==19.0.0
fonttools==4.40.0
kiwisolver==1.4.4
matplotlib==3.7.2
numpy==1.25.1
openpyxl==3.0.10
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pyparsing==3.0.9
PyQt6==6.5.1
PyQt6-Qt6==6.5.1
PyQt6-sip==13.5.1
python-dateutil==2.8.2
pytz==2023.3
rstr==3.2.1
six==1.16.0
tzdata==2023.3
